♻️ Refactor : 개별피드 기본구조 컴포넌트로 변경
♻️  refactor : ProfileSection 컴포넌트화
