+++ src/components/question/FloatingButton.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function FloatingButton() {
+  return (
+    <div className="fixed bottom-[24px] right-[24px]">
+      <button className="px-[49.5px] py-[14.5px] bg-[#542f1a] rounded-full">
+        질문 작성하기
+      </button>
+    </div>
+  );
+}
+
+export default FloatingButton;
